entityManagerFactory

+++ Employee_Batch_Processing_Demo/src/main/java/com/example/demo/repository/StudentRepo.java
@@ -11,6 +11,6 @@ import com.example.demo.entity.StudentEntity;
  */
 
 @Repository
-public interface StudentRepo extends JpaRepository<StudentEntity, Long> {
+public interface StudentRepo extends JpaRepository<StudentEntity, Integer> {
 
 }
