+++ Employee_Batch_Processing_Demo/src/main/java/com/example/demo/service/StudentService.java
@@ -15,7 +15,7 @@ public interface StudentService {
 	
 	void save(StudentEntity studentEntity);
 	List<StudentEntity> getStudents();
-	StudentEntity getStudentById(Long stId);
+	StudentEntity getStudentById(Integer stId);
 	
 
 }
